requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.11.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
django-extensions>=3.1.5
